requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
faiss-cpu
sentence-transformers
langchain-community
google-generativeai
